Move CurrencySelectorElement to elements package
Committed-By-Agent: codex
